DJANGO/250616/requirements.txt
django-environ

pyhub-llm[all]

# 모든 python 애플리케이션 저장소에는 반드시 있는 파일
# 해당 애플리케이션 구동에 필요한 모든 라이브러리들을 1줄에 하나씩 명시
# 한방 설치 : pip install -r requirements.txt
# 라이브러리 의존성 명시
# 의존성(dependencies)
# 이 애플리케이션 구동에 꼭 필요하다!라는 뜻
# 즉 이 애플리케이션을 구동하는데 필요한 라이브러리를 명시해두는 것

django~=5.2.0

# major, minor, bugfix 버전이라는게 존재한다
# bugfix 버전은 기능이 변경되진 않고 버그만 변경된다, 마지막 숫자를 올린다

# == : 정확히 같은 버전 설치
# ~= : bugfix버전 최신을 설치 만약 5.2.3버전이 있다면 5.2.0을 써도 5.2.3 설치
# <= : 해당 버전에서 가장 최저버전 설치
# >= : 해당 버전보다 최신 버전 설치


